#2582: fix unregistering quickbar actions
