+++ src/common.ts
@@ -23,8 +23,8 @@ import {
   SEARCH_WINDOW,
 } from "@/messaging/constants";
 
-export const NOFICATIONS_Z_INDEX = 2_147_483_647;
-export const MAX_Z_INDEX = NOFICATIONS_Z_INDEX - 1; // Let notifications always be higher
+export const NOTIFICATIONS_Z_INDEX = 2_147_483_647;
+export const MAX_Z_INDEX = NOTIFICATIONS_Z_INDEX - 1; // Let notifications always be higher
 export const PANEL_FRAME_ID = "pixiebrix-extension";
 export const PIXIEBRIX_DATA_ATTR = "data-pb-uuid";
 export const PIXIEBRIX_READY_ATTRIBUTE = "data-pb-ready";
