+++ src/components/quickBar/QuickBarApp.tsx
@@ -15,19 +15,16 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import React, { forwardRef, useEffect } from "react";
+import React, { useEffect } from "react";
 import ReactDOM from "react-dom";
 import {
   Action,
-  ActionId,
-  ActionImpl,
   KBarAnimator,
   KBarPositioner,
   KBarProvider,
-  KBarResults,
   KBarSearch,
   useKBar,
-  useMatches,
+  useRegisterActions,
   VisualState,
 } from "kbar";
 import ReactShadowRoot from "react-shadow-root";
@@ -35,10 +32,12 @@ import quickBarRegistry from "@/components/quickBar/quickBarRegistry";
 import faStyleSheet from "@fortawesome/fontawesome-svg-core/styles.css?loadAsUrl";
 import { expectContext } from "@/utils/expectContext";
 import { once } from "lodash";
-import { NOFICATIONS_Z_INDEX } from "@/common";
+import { NOTIFICATIONS_Z_INDEX } from "@/common";
 import { useEventListener } from "@/hooks/useEventListener";
 import { Stylesheet } from "@/components/Stylesheet";
 import selection from "@/utils/selectionController";
+import { animatorStyle, searchStyle } from "./quickBarTheme";
+import QuickBarResults from "./QuickBarResults";
 
 /**
  * Set to true if the KBar should be displayed on initial mount (i.e., because it was triggered by the
@@ -47,171 +46,9 @@ import selection from "@/utils/selectionController";
 let autoShow = false;
 
 /**
- * Window event name to programmatically trigger quickbar
+ * Window event name to programmatically trigger quick bar
  */
 const QUICKBAR_EVENT_NAME = "pixiebrix-quickbar";
-
-const theme = {
-  background: "rgb(252, 252, 252)",
-  foreground: "rgb(28, 28, 29)",
-  shadow: "0px 6px 20px rgba(0, 0, 0, 20%)",
-  a1: "rgba(0, 0, 0, 0.05)",
-  a2: "rgba(0, 0, 0, 0.1)",
-};
-
-const searchStyle = {
-  padding: "12px 16px",
-  fontSize: "16px",
-  width: "100%",
-  boxSizing: "border-box" as React.CSSProperties["boxSizing"],
-  outline: "none",
-  border: "none",
-  background: theme.background,
-  color: theme.foreground,
-};
-
-const animatorStyle: React.CSSProperties = {
-  all: "initial",
-  fontFamily: "sans-serif",
-  maxWidth: "600px",
-  width: "100%",
-  background: theme.background,
-  color: theme.foreground,
-  borderRadius: "8px",
-  overflow: "hidden",
-  boxShadow: theme.shadow,
-};
-
-const groupNameStyle = {
-  padding: "8px 16px",
-  fontSize: "10px",
-  textTransform: "uppercase" as const,
-  opacity: 0.5,
-};
-
-const ResultItem = forwardRef(
-  (
-    {
-      action,
-      active,
-      currentRootActionId,
-    }: {
-      action: ActionImpl;
-      active: boolean;
-      currentRootActionId: ActionId;
-    },
-    ref: React.Ref<HTMLDivElement>
-  ) => {
-    const ancestors = React.useMemo(() => {
-      if (!currentRootActionId) return action.ancestors;
-      const index = action.ancestors.findIndex(
-        (ancestor) => ancestor.id === currentRootActionId
-      );
-      // +1 removes the currentRootAction; e.g.
-      // if we are on the "Set theme" parent action,
-      // the UI should not display "Set theme… > Dark"
-      // but rather just "Dark"
-      return action.ancestors.slice(index + 1);
-    }, [action.ancestors, currentRootActionId]);
-
-    return (
-      <div
-        ref={ref}
-        style={{
-          padding: "12px 16px",
-          background: active ? theme.a1 : "transparent",
-          borderLeft: `2px solid ${active ? theme.foreground : "transparent"}`,
-          display: "flex",
-          alignItems: "center",
-          justifyContent: "space-between",
-          cursor: "pointer",
-        }}
-      >
-        <div
-          style={{
-            display: "flex",
-            gap: "8px",
-            alignItems: "center",
-            fontSize: 14,
-          }}
-        >
-          {action.icon && action.icon}
-          <div style={{ display: "flex", flexDirection: "column" }}>
-            <div>
-              {ancestors.length > 0 &&
-                ancestors.map((ancestor) => (
-                  <React.Fragment key={ancestor.id}>
-                    <span
-                      style={{
-                        opacity: 0.5,
-                        marginRight: 8,
-                      }}
-                    >
-                      {ancestor.name}
-                    </span>
-                    <span
-                      style={{
-                        marginRight: 8,
-                      }}
-                    >
-                      &rsaquo;
-                    </span>
-                  </React.Fragment>
-                ))}
-              <span>{action.name}</span>
-            </div>
-            {action.subtitle && (
-              <span style={{ fontSize: 12 }}>{action.subtitle}</span>
-            )}
-          </div>
-        </div>
-        {action.shortcut?.length ? (
-          <div
-            aria-hidden
-            style={{ display: "grid", gridAutoFlow: "column", gap: "4px" }}
-          >
-            {action.shortcut.map((sc: string) => (
-              <kbd
-                key={sc}
-                style={{
-                  padding: "4px 6px",
-                  background: "rgba(0 0 0 / .1)",
-                  borderRadius: "4px",
-                  fontSize: 14,
-                }}
-              >
-                {sc}
-              </kbd>
-            ))}
-          </div>
-        ) : null}
-      </div>
-    );
-  }
-);
-
-ResultItem.displayName = "ResultItem";
-
-const RenderResults: React.FC = () => {
-  const { results, rootActionId } = useMatches();
-
-  return (
-    <KBarResults
-      items={results}
-      onRender={({ item, active }) =>
-        typeof item === "string" ? (
-          <div style={groupNameStyle}>{item}</div>
-        ) : (
-          <ResultItem
-            action={item}
-            active={active}
-            currentRootActionId={rootActionId}
-          />
-        )
-      }
-    />
-  );
-};
 
 // Modeled around KBarPortal https://github.com/timc1/kbar/blob/a232f3d8976a61eeeb844152dea25a23a76ad368/src/KBarPortal.tsx
 const KBarToggle: React.FC = (props) => {
@@ -235,18 +72,28 @@ const KBarToggle: React.FC = (props) => {
 };
 
 function useActions(): void {
+  // The useActions hook is included in KBarComponent, which mounts/unmounts when the kbar is toggled
+
+  // The kbar useRegisterActions hook uses an "unregister" affordance that's not available in the types
+  // https://github.com/timc1/kbar/blob/main/src/useStore.tsx#L63
+  // https://github.com/timc1/kbar/blob/main/src/useRegisterActions.tsx#L19
+  useRegisterActions(quickBarRegistry.currentActions, []);
+
+  // Use the query directly for updating while the page editor is open. The useRegisterActions hook doesn't seem to
+  // work for that 🤷 even if actions are in the dependency list
   const { query } = useKBar();
 
+  // Listen for changes while the kbar is mounted (e.g., the user is making edits in the page editor)
   useEffect(() => {
-    const handler = (actions: Action[]) => {
-      query.registerActions(actions);
+    const handler = (nextActions: Action[]) => {
+      query.registerActions(nextActions);
     };
 
     quickBarRegistry.addListener(handler);
     return () => {
       quickBarRegistry.removeListener(handler);
     };
-    // eslint-disable-next-line react-hooks/exhaustive-deps -- only run on mount; query will be defined on initial mount
+    // eslint-disable-next-line react-hooks/exhaustive-deps -- the query is available on initial mount
   }, []);
 }
 
@@ -273,14 +120,16 @@ const AutoShow: React.FC = () => {
 };
 
 const KBarComponent: React.FC = () => {
+  // The KBarComponent is mounted/unmounted on kbar toggle (controlled by KBarToggle)
+
   useActions();
 
   return (
     // Like notifications, this is temporary UI that must appear above everything
-    <KBarPositioner style={{ zIndex: NOFICATIONS_Z_INDEX }}>
+    <KBarPositioner style={{ zIndex: NOTIFICATIONS_Z_INDEX }}>
       <KBarAnimator style={animatorStyle}>
         <KBarSearch style={searchStyle} />
-        <RenderResults />
+        <QuickBarResults />
       </KBarAnimator>
     </KBarPositioner>
   );
