+++ src/contentScript/notify.tsx
@@ -20,7 +20,7 @@ import { render } from "react-dom";
 import { toast, Toaster } from "react-hot-toast";
 import { uuidv4 } from "@/types/helpers";
 import { DefaultToastOptions } from "react-hot-toast/dist/core/types";
-import { NOFICATIONS_Z_INDEX } from "@/common";
+import { NOTIFICATIONS_Z_INDEX } from "@/common";
 
 type NotificationType = "info" | "success" | "error" | "loading";
 interface Notification {
@@ -31,7 +31,7 @@ interface Notification {
 }
 
 const containerStyle: React.CSSProperties = {
-  zIndex: NOFICATIONS_Z_INDEX,
+  zIndex: NOTIFICATIONS_Z_INDEX,
   fontFamily: "sans-serif",
 };
